refactor(tui): colocate Screen::Revisions into tui/screens/revisions.rs
Eighth screen migration for #287 (Phase 2). Moved into the new
tui/screens/revisions.rs: handle_key_revisions, revisions_guard (the
revision_diff_incremental_intent, revision_diff_intent,
restore_revision_preview_intent, revision_target_file_label (all
Revisions-exclusive), revision_row_label (moved from render.rs, only
called by build_revisions_vm), build_revisions_vm, render_revisions_vm,
revisions_palette_items.

Shared-helper notes: revision_version_label (bg.rs) stays put, already
reachable via its existing pub(super); diff_allows_sync and group_by_id
stay in mod.rs (general-purpose, used elsewhere). revision_target_file_label
bumped to pub(crate) (tests.rs calls it directly on AppState).
HelpTopic::for_screen's Revisions arm is a plain variant reference with
no logic to move, so it stays as-is (matches the Pins/Gists/GistDetail
arms, not the Config/Help HELP_TOPIC const pattern).

Pure move — bodies and existing tests unchanged.

Part of #287
